Move store and lockfiles behind internal package.
