+++ internal/codeintel/dependencies/internal/lockfiles/errors.go
