+++ internal/codeintel/dependencies/internal/lockfiles/gen.go
@@ -1,0 +1,3 @@
+package lockfiles
+
+//go:generate ../../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/internal/codeintel/dependencies/internal/lockfiles -i GitService -o mock_iface_test.go
