+++ internal/codeintel/dependencies/internal/lockfiles/go.go
