+++ internal/codeintel/dependencies/internal/lockfiles/iface.go
