+++ internal/codeintel/dependencies/internal/lockfiles/init.go
