+++ internal/codeintel/dependencies/internal/lockfiles/npm.go
