+++ internal/codeintel/dependencies/internal/lockfiles/observability.go
