+++ internal/codeintel/dependencies/internal/lockfiles/parser.go
