+++ internal/codeintel/dependencies/internal/lockfiles/parser_test.go
