+++ internal/codeintel/dependencies/internal/lockfiles/service.go
