+++ internal/codeintel/dependencies/internal/lockfiles/service_test.go
