+++ internal/codeintel/dependencies/internal/lockfiles/testdata/parse/go.mod/go_real_world.mod.golden
