+++ internal/codeintel/dependencies/internal/lockfiles/testdata/parse/go.mod/go_real_world_with_replace_and_exclude.mod.golden
