+++ internal/codeintel/dependencies/internal/lockfiles/testdata/parse/yarn.lock/yarn_normal.lock.golden
