+++ internal/codeintel/dependencies/internal/lockfiles/testdata/parse/yarn.lock/yarn_v2_normal.lock
