+++ internal/codeintel/dependencies/internal/lockfiles/testdata/parse/yarn.lock/yarn_with_npm.lock.golden
