+++ internal/codeintel/dependencies/internal/lockfiles/testdata/svc/TestListDependencies/go.golden
