+++ internal/codeintel/dependencies/internal/lockfiles/testdata/svc/TestListDependencies/npm.golden
