+++ internal/codeintel/dependencies/internal/store/consts.go
