+++ internal/codeintel/dependencies/internal/store/observability.go
