+++ internal/codeintel/dependencies/internal/store/scan.go
