--- internal/codeintel/dependencies/lockfiles/gen.go
@@ -1,3 +1,0 @@
-package lockfiles
-
-//go:generate ../../../../dev/mockgen.sh github.com/sourcegraph/sourcegraph/internal/codeintel/dependencies/lockfiles -i GitService -o mock_iface_test.go
